If updating HPA fails due to conflict, try again up to 3 times.
This is a common concurrency case - when doing get, update, put on a resource
which may be updated elsewhere, one has to either use synchronization mechanisms
or retry.

I don't believe that k8s has a synchronization mechanism to prevent against
the HPA metrics checking from updating the status at exactly the same time
that ss is trying to update the min/max replicas.

Since this new code is retrying, there _is_ a risk in colliding with another
ss cron entry as well, and there _are_ synchronization mechanisms (mutex)
to help ensure that the currently scheduling thing (which got there first)
gets to retry, before the new one comes in.

Fixes #28.